Project/PDFTalker/requirements.txt
streamlit==1.26.0
langchain
langchain-chroma
langchain-google-genai
langchain-community
python-dotenv
pypdf
python-docx
